// src/types/cssClassEnums.ts
export enum BUTTON_PARTS {
  BUTTON = 'button',
  BUTTON_DISABLED = 'button-disabled',
  BUTTON_OUTLINED = 'button-outlined',
  BUTTON_OUTLINED_FOCUSED = 'button-outlined-focused',
  BUTTON_OUTLINED_DISABLED = 'button-outlined-disabled',
  BUTTON_START_ICON = 'button-start-icon',
  BUTTON_START_ICON_SMALL_WITH_PADDING = 'button-start-icon-small-with-padding',
  BUTTON_START_ICON_SMALL_WITHOUT_PADDING = 'button-start-icon-small-without-padding',
  BUTTON_START_ICON_MEDIUM_WITH_PADDING = 'button-start-icon-medium-with-padding',
  BUTTON_START_ICON_MEDIUM_WITHOUT_PADDING = 'button-start-icon-medium-without-padding',
  BUTTON_START_ICON_FAB = 'button-start-icon-fab',
  BUTTON_END_ICON = 'button-end-icon',
  BUTTON_START_ICON_RTL_MARGIN = 'button-start-icon-rtl-margin',
  BUTTON_TEXT = 'button-text',
  BUTTON_CONTAINED = 'button-contained',
  BUTTON_CONTAINED_INVERSE = 'button-contained-inverse',
  BUTTON_CONTAINED_DISABLED = 'button-contained-disabled',
  BUTTON_CONTAINED_DISABLED_INVERSE = 'button-contained-disabled-inverse',
  BUTTON_CONTAINED_HOVERED = 'button-contained-hovered',
  BUTTON_CONTAINED_FOCUSED = 'button-contained-focused',
  BUTTON_CONTAINED_FOCUSED_INVERSE = 'button-contained-focused-inverse',
  BUTTON_CONTAINED_HOVERED_FOCUSED = 'button-contained-hovered-and-focused',
  BUTTON_ENCHANTED_TEXT = 'button-enchanted-text',
  BUTTON_ENCHANTED_TEXT_INVERSE = 'button-enchanted-text-inverse',
  BUTTON_ENCHANTED_TEXT_DISABLED = 'button-enchanted-text-disabled',
  BUTTON_ENCHANTED_TEXT_DISABLED_INVERSE = 'button-enchanted-text-disabled-inverse',
  BUTTON_ENCHANTED_TEXT_HOVERED = 'button-enchanted-text-hovered',
  BUTTON_ENCHANTED_TEXT_FOCUSED = 'button-enchanted-text-focused',
  BUTTON_ENCHANTED_TEXT_FOCUSED_INVERSE = 'button-enchanted-text-focused-inverse',
  BUTTON_ENCHANTED_TEXT_HOVERED_FOCUSED = 'button-enchanted-text-hovered-and-focused',
  BUTTON_ENCHANTED_OUTLINED = 'button-enchanted-outlined',
  BUTTON_ENCHANTED_OUTLINED_FOCUSED = 'button-enchanted-outlined-focused',
  BUTTON_ENCHANTED_OUTLINED_DISABLED = 'button-enchanted-outlined-disabled',
  BUTTON_ENCHANTED_OUTLINED_HOVERED = 'button-enchanted-outlined-hovered',
  BUTTON_ENCHANTED_OUTLINED_HOVERED_FOCUSED = 'button-enchanted-outlined-hovered-and-focused',
}
// ...
export enum HEADER_VARIANT {
  HEADER_AUTHORING = 'header-authoring-modal',
  HEADER_AUTHORING_MODAL_CLOSE = 'header-authoring',
  HEADER_ENDUSER = 'header-endUser',
}

export enum HEADER_PARTS {
  BADGE_TEXT = 'badge_text',
  BADGE_TEXT_RTL = 'badge_text-rtl',
  BADGE_DOT = 'badge_dot',
  H6 = 'h6',
  INPUT = 'input',
  HEADER = 'header',
  SUB_HEADER_START = 'sub-header-start',
  SUB_HEADER_END = 'sub-header-end',
  HEADER_SPACING_START = 'header-spacing-start',
  HEADER_SPACING_START_HAMBURGER = 'header-spacing-start-hamburger',
  HEADER_SPACING_END = 'header-spacing-end',
  HR_PART = 'hr-part',
}